Spawn Headphones into players real world.
Moved away from the LWRP due to some incompatabilities.

Wrote a simple script that uses the users tap to spawn the headphones prefab.

Tasks for tomorrow will be to add interaction features which will involve users being able to tap icons on the model to learn more about the headphones, alongside changing size and rotation.