Built Object and UI managers to allow for interaction
Created a simple slider to allow the AR object to be rotated (this is being controlled by the AR Object itself in te update function)

I also added a video texture to a quad which can be activated through a simple button and hidden in a simliar way, a future expansion would be to have the quad always facing the camera.

next up is to create the information points on the model and add some of the data.