First Release
This first release is a very early example of what AR can achive for advertising.

The instructions at the beginning last 5 seconds to give the user time to read and udnertand.

once a surface is scanned and tapped the AR Object in placed, this object can be roatated with the slider.

the 2 buttons along the botom of the slider will play the promo video above the AR object (can be closed with the cross) and a link to the website (though I want to look at adding a check to ensure the user is happy being sent away from my application).

around the headset are 3 'feature' icons that can be pressed to give you some of the features of the headphones (this could easily be expanded upon over time as well with more information)

//KNOWN BUGS

- the top 'feature' button cannot be pressed from behind. A potential fix for this is to always have it facing the camera.

//Future Improvments

-I would like to stop the room scan once the object is placed, I suspect this will be easy to do.

-Use AR Kit's face mapping system to allow the user to 'try on' the headphones before they buy (a little like a snapchat filter) this feature only works on AR Kit so would not be avalible on android and at the point of writing I dont have an Apple device for testing or building.